PythonSerialController/requirements.txt
keyboard==0.13.5
pygame==2.0.1
pyserial==3.5
